+++ src/jdksavdecc_descriptor_storage.c
@@ -163,7 +163,7 @@ uint16_t jdksavdecc_descriptor_storage_buffer_read_descriptor(
         uint16_t configuration_number,
         uint16_t descriptor_type,
         uint16_t descriptor_index,
-        uint16_t *result_buffer,
+        uint8_t *result_buffer,
         uint16_t result_buffer_len) {
     uint16_t r=0;
     struct jdksavdecc_descriptor_storage *self = (struct jdksavdecc_descriptor_storage *)self_;
